requirements.txt
pytorch
nltk
requests
ujson
google-cloud
python-socketio
eventlet
pattern
nltk
